ft: Several changes for informing documents, as well as to chat with the AI
