+++ src/app/opos/page.tsx
@@ -8,16 +8,21 @@ import { data, userData } from "../store";
 import { useAtom } from "jotai";
 import Markdown from "markdown-to-jsx";
 import { Bars } from "react-loader-spinner";
+import { bagsCoins, bagsNfts } from "../store/bags";
 
 export default function OPOS() {
   const lastBotMessageIndex = React.useRef(0);
   const [currentUser] = useAtom(data);
   const [user] = useAtom(userData);
+  const [nfts] = useAtom(bagsNfts);
+  const [coins] = useAtom(bagsCoins)
 
   const [text, setText] = React.useState("");
   const [messages, setMessages] = React.useState<AIChatMessage[]>([]);
   const [isDisabled, setIsDisabled] = React.useState(false);
-  const [isLoading, setIsLoading] = React.useState(false);
+  const [isLoading, setIsLoading] = React.useState(false); 
+
+  const [availableNamespaces, setAvailableNamespaces] = React.useState<string[]>([]);
 
   const messagesRef = React.useRef<HTMLDivElement>(null);
 
@@ -44,7 +49,7 @@ export default function OPOS() {
           body: JSON.stringify({
             username: currentUser?.profile.username ?? user?.name ?? "Visitor",
             prompt: text,
-            namespaces: [],
+            namespaces: availableNamespaces,
           }),
         },
       );
@@ -142,6 +147,20 @@ export default function OPOS() {
     },
     [currentUser, user],
   );
+
+  React.useEffect(() => {
+    const namespaces: string[] = [];
+
+    coins?.memecoins.forEach((coin) => namespaces.push(coin.symbol));
+    coins?.community.forEach((coin) => namespaces.push(coin.symbol));
+
+    nfts?.badges.forEach((coin) => namespaces.push(coin.symbol));
+    nfts?.passes.forEach((coin) => namespaces.push(coin.symbol));
+    nfts?.profiles.forEach((coin) => namespaces.push(coin.symbol));
+
+    setAvailableNamespaces(namespaces);
+    
+  }, [coins, nfts]);
 
   return (
     <div className="background-content flex w-full h-full justify-center">
